src/test/java/ru/job4j/concurrent/synchronize/SingleLockListTest.java
package ru.job4j.concurrent.synchronize;

import org.junit.Test;
import java.util.Set;
import java.util.TreeSet;
import static org.hamcrest.core.Is.is;
import static org.junit.Assert.*;

public class SingleLockListTest {

    @Test
    public void add() throws InterruptedException {
        SingleLockList<Integer> list = new SingleLockList<>();
        Thread first = new Thread(() -> list.add(1));
        Thread second = new Thread(() -> list.add(2));
        first.start();
        second.start();
        first.join();
        second.join();
        Set<Integer> rsl = new TreeSet<>();
        list.iterator().forEachRemaining(rsl::add);
        assertThat(rsl, is(Set.of(1, 2)));
    }

    @Test
    public void addMultiple() throws InterruptedException {
        SingleLockList<Integer> list = new SingleLockList<>();
        Thread first = new Thread(
                () -> {
                    for (int i = 0; i < 10000; i++) {
                        list.add(i);
                    }
                });
        Thread second = new Thread(
                () -> {
                    for (int i = 10000; i < 20000; i++) {
                        list.add(i);
                    }
                });
        first.start();
        second.start();
        first.join();
        second.join();
        Set<Integer> rsl = new TreeSet<>();
        list.iterator().forEachRemaining(rsl::add);
        assertThat(rsl.size(), is(20000));
    }
}